+++ ledger/participant-integration-api/src/main/scala/platform/config/MetricsConfig.scala
@@ -1,13 +1,12 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.runner.common
+package com.daml.platform.config
 
-import com.daml.ledger.runner.common.MetricsConfig.MetricRegistryType
 import com.daml.metrics.MetricsReporter
+import com.daml.platform.config.MetricsConfig.MetricRegistryType
 
-import scala.concurrent.duration.Duration
-import scala.concurrent.duration._
+import scala.concurrent.duration.{Duration, _}
 
 final case class MetricsConfig(
     enabled: Boolean = false,
