+++ ledger/participant-integration-api/src/main/scala/platform/config/ParticipantRunMode.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.runner.common
+package com.daml.platform.config
 
 sealed abstract class ParticipantRunMode
